requirements.txt
fastapi
uvicorn
requests
pydantic
langchain
pypdf
faiss-cpu
python-dotenv
groq
